backend/requirements.txt
Django==6.0.1
djangorestframework
django-cors-headers
gunicorn
whitenoise
Pillow